requirements.txt
streamlit
pandas
matplotlib
ibm-watsonx-ai
python-dotenv
